fix: various cleanup after content-publishing config refactor
