+++ apps/content-publishing-api/src/swagger_config.ts
