+++ libs/content-publishing-lib/src/config/index.ts
@@ -1,2 +1,1 @@
 export * from './ipfs.config';
-export * from './swagger_config';
